% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{create_tree}
\alias{create_tree}
\title{Create a persistent expandable k-d tree}
\usage{
create_tree(x, leafSize = 11, leafBuffer = 0)
}
\arguments{
\item{x}{n vectors stored in an n by p matrix.  These vectors will be used
to create the K-d tree.  It is assumed that all values are finite and 
observed.}

\item{leafSize}{A scalar used to specify the number of points to store in the 
leaf nodes.}

\item{leafBuffer}{A scalar used to specify the number of empty index locations
to retrain in the leaf nodes.  This is useful to allow for quicker insertions
of new data points.  Note that empty points are in addition to the leafSize.}
}
\value{
An S3 object cotaining the persistent tree.
}
\description{
\code{create_tree} build a persistent expandable k-d tree from a vector or 
matrix.
}
\examples{
x <- matrix(runif(20),10,2)
myTree <- create_tree(x)
}

