% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_authors.R
\name{get_authors}
\alias{get_authors}
\title{Get author metadata}
\usage{
get_authors(corpus = "cs")
}
\arguments{
\item{corpus}{string, ISO code of the corpus language. Check get_metadata() for all available corpora.}
}
\value{
returns tidy data frame (tibble)
}
\description{
Get author metadata
}
\examples{
\dontrun{
df <- get_authors(corpus="en")
}
}
