% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{List all PoeTree corpora}
\usage{
get_metadata()
}
\value{
returns tidy data frame (tibble)
}
\description{
List all PoeTree corpora
}
\examples{
\dontrun{
df <- get_metadata()
}
}
