% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_poems.R
\name{get_poems}
\alias{get_poems}
\title{List PoeTree poems (metadata only)}
\usage{
get_poems(
  corpus = "cs",
  authors = NULL,
  duplicates = FALSE,
  tidy = TRUE,
  list_all = FALSE
)
}
\arguments{
\item{corpus}{string, ISO code of the corpus language. Check get_metadata() for all available corpora.}

\item{authors}{integer or vector of integers, author ID. Check get_authors() for these. Supports multiple ids provided in a vector.}

\item{duplicates}{logical. Whether to include duplicated texts. Default is FALSE. See more about duplicate detection strategy: https://versologie.cz/poetree/deduplication/}

\item{tidy}{logical. If TRUE it unnests poem records that have multiple authors (one poem-author per row format). If FALSE, it bundles author IDs in a string (one poem per row format). Very rarely useful, TRUE by default}

\item{list_all}{logical. If TRUE, lists all poems available in the corpus. Default is FALSE.}
}
\value{
returns tidy data frame (tibble)
}
\description{
List PoeTree poems (metadata only)
}
\examples{
\dontrun{
df <- get_poems(corpus="cs",author=c(81,52))
}
}
