% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_poems_a.R
\name{get_poems_a}
\alias{get_poems_a}
\title{Get poem metadata by author ID}
\usage{
get_poems_a(corpus, author, duplicates = FALSE)
}
\arguments{
\item{corpus}{string, ISO code of the corpus language. Check get_metadata() for all available corpora.}

\item{author}{integer, author ID. Check get_authors() for these.}

\item{duplicates}{logical. Whether to include duplicated texts. Default is FALSE. See more about duplicate detection strategy: https://versologie.cz/poetree/deduplication/}
}
\value{
returns tidy data frame (tibble)
}
\description{
Get poem metadata by author ID
}
\examples{
\dontrun{
df <- get_poems_a(corpus="cs",author=81)
}
}
