% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text.R
\name{get_text}
\alias{get_text}
\title{Retrieve text of a poem}
\usage{
get_text(corpus, poem_id, output = "lines", metadata = FALSE)
}
\arguments{
\item{corpus}{string, ISO code of the corpus language. Check get_metadata() for all available corpora.}

\item{poem_id}{integer, corpus-wide poem ID. Check get_poems() for these.}

\item{output}{string. Controls how extensive should be the output. Options are "lines" (one line = one row), "tokenized" (one token = one row) and "annotated"}

\item{metadata}{logical. Whether to include poems' metadata. Default is FALSE.}
}
\value{
returns tidy data frame (tibble)
}
\description{
Retrieve text of a poem
}
\examples{
\dontrun{
df <- get_text(corpus="cs",poem_id=1)
}
}
