% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_tree.R
\name{search_tree}
\alias{search_tree}
\title{Search for k nearsest neighbors in a persistent expandable k-d tree}
\usage{
search_tree(y, tree, k = 5)
}
\arguments{
\item{y}{A vector to search for in the k-d tree.}

\item{tree}{A K-d tree stored as a POETree object.}

\item{k}{A scalar used to specify the number of neighbors to return from the 
nearest neighbor search.}
}
\value{
A vector of integers identifying nearest neighbors.
}
\description{
\code{search_tree} build a persistent expandable k-d tree from a vector or 
matrix.
}
\examples{
x <- matrix(runif(20),10,2)
# find the four closest nearest neighbors
myTree <- create_tree(x)
y <- c(0,0)
neighbors <- search_tree(y,myTree,k=4)
}

