% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_authors.R
\name{get_authors}
\alias{get_authors}
\title{Get author metadata}
\usage{
get_authors(x = "../samples_json/authors.json")
}
\arguments{
\item{x}{json file}
}
\value{
returns tidy data frame (tibble)
}
\description{
Get author metadata
}
\examples{
\dontrun{
df <- get_authors()
}
}
